Bugfixing in the movegen code
Changed the jump move generation code to use indexes instead of iterators for iterating over the growing movelist because I needed to dynamically add and remove elements from the movelist while iterating. It took me a while (read: a very long time) to figure out that it was the call to push_back, not the call to erase, that was invalidating iterators and causing trouble. Also made some other bugfixes in the movegen code and added file types generated by building with debugging support to .gitignore.