The simple AIs can now be played against.
Added support for choosing the opponent that one wishes to play against, which at the moment can be oneself or one of the simple AIs. The support for this is pretty hardcoded and hackish; I'll revise all of this once I have a more solid AI subsystem going. The goal of this commit is just to get some AIs going to make testing stuff in the checkers subsystem easier and more fun.

For some reason, the calls to makeMove with the AI-selected moves do not always result in a completed move. Todo: investigate this.