Added the option to play ais against each other
Added the option to play two ais against each other for a given number of games and see the full game (in the case of one game selected, currently implemented) or a summary of some statistics related to the games such as wins, losses, and so on (in the case of multiple games selected, currently not implemented).

The ais can't get through one game without one of them coming up with a move that can't be played...not their fault though, the checkers movegen stuff is probably buggy as hell.